Fix init CNode depth conversions

+++ apps/root-task/src/bootstrap/cspace_sys.rs
@@ -146,7 +146,7 @@ pub fn cnode_copy_raw_single(
     src_root: sys::seL4_CNode,
     src_slot: sys::seL4_Word,
 ) -> sys::seL4_Error {
-    let depth_bits = sel4::init_cnode_bits(bi);
+    let depth_bits = usize::from(sel4::init_cnode_bits(bi));
     let depth = depth_bits as sys::seL4_Word;
     let depth_u8 = bits_as_u8(depth_bits);
     let rights = sys::seL4_CapRights::new(1, 1, 1, 1);
